fix: background image upload
